changes in q1

+++ src/main/scala/com/assignment/models/ScoreCard.scala
@@ -1,3 +1,3 @@
 package com.assignment.models
 
-case class ScoreCard(studentId: Long, marks: Map[Long, Float], percentage: Float)
+case class ScoreCard(studentId: Long, marks: Map[Int, Int], percentage: Float)
